chore: Update file structure and add mock files for tasks 1-4
This commit updates the file structure of the project and adds mock files for tasks 1-4 in the dashboard directory. The changes include adding fileMock.js and styleMock.js files for each task, as well as updating the index.js file in each task directory to import the App component and render it in the root element. Additionally, the CourseShape.js file in each task's CourseList directory has been updated to include prop types for id, name, and credit. Finally, the Footer component has been updated to import the getFullYear and getFooterCopy functions from the utils/utils module and display the copyright information in the footer element.